Initial Client-Server
The server is in the classdemo-master folder. Open ClientServer.sln and
run only the server project (ignore the client project).

The Unity file didn't go through any changes other than attaching the
Client.cs script to the Player object. The Client.cs file has a very
strange bug, i.e. messages won't send unless there is a Debug.Log()
there. Not sure why.

As of this build, the client is able to connect to the server and send
it an initial message. The server receives the message successfully.
However the client is unable to receive the server's echoing of the
client's initial message. It seems that the client is able to receive
it, but the processing of the message gets cut-off and dropped at some
point. I/we will have to look into this.

<EOF> ok just kidding